Refactor UDFs into their own package
